chore: 예외처리 클래스 도메인별로 패키지 분리
