+++ src/main/java/com/example/tnote/base/exception/common/CommonException.java
@@ -1,5 +1,6 @@
-package com.example.tnote.base.exception;
+package com.example.tnote.base.exception.common;
 
+import com.example.tnote.base.exception.common.CommonErrorResult;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
