+++ src/main/java/com/example/tnote/base/exception/jwt/JwtException.java
@@ -1,5 +1,6 @@
-package com.example.tnote.base.exception;
+package com.example.tnote.base.exception.jwt;
 
+import com.example.tnote.base.exception.jwt.JwtErrorResult;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
