+++ src/main/java/com/example/tnote/base/exception/todo/TodoErrorResult.java
@@ -1,4 +1,4 @@
-package com.example.tnote.base.exception;
+package com.example.tnote.base.exception.todo;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
