+++ src/main/java/com/example/tnote/base/exception/todo/TodoException.java
@@ -1,5 +1,6 @@
-package com.example.tnote.base.exception;
+package com.example.tnote.base.exception.todo;
 
+import com.example.tnote.base.exception.todo.TodoErrorResult;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
