+++ src/main/java/com/example/tnote/base/exception/user/UserException.java
@@ -1,5 +1,6 @@
-package com.example.tnote.base.exception;
+package com.example.tnote.base.exception.user;
 
+import com.example.tnote.base.exception.user.UserErrorResult;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
