+++ src/main/java/com/example/tnote/base/filter/JwtAuthenticationFilter.java
@@ -1,7 +1,6 @@
 package com.example.tnote.base.filter;
 
-import com.example.tnote.base.exception.JwtErrorResult;
-import com.example.tnote.base.exception.JwtException;
+import com.example.tnote.base.exception.jwt.JwtException;
 import com.example.tnote.base.utils.JwtTokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
