+++ src/main/java/com/example/tnote/boundedContext/RefreshToken/service/RefreshTokenService.java
@@ -1,9 +1,9 @@
 package com.example.tnote.boundedContext.RefreshToken.service;
 
-import com.example.tnote.base.exception.CommonErrorResult;
-import com.example.tnote.base.exception.CommonException;
-import com.example.tnote.base.exception.JwtErrorResult;
-import com.example.tnote.base.exception.JwtException;
+import com.example.tnote.base.exception.common.CommonErrorResult;
+import com.example.tnote.base.exception.common.CommonException;
+import com.example.tnote.base.exception.jwt.JwtErrorResult;
+import com.example.tnote.base.exception.jwt.JwtException;
 import com.example.tnote.boundedContext.RefreshToken.entity.RefreshToken;
 import com.example.tnote.boundedContext.RefreshToken.repository.RefreshTokenRepository;
 import com.example.tnote.boundedContext.user.repository.UserRepository;
