+++ src/main/java/com/example/tnote/boundedContext/consultation/service/ConsultationService.java
@@ -1,7 +1,7 @@
 package com.example.tnote.boundedContext.consultation.service;
 
-import com.example.tnote.base.exception.UserErrorResult;
-import com.example.tnote.base.exception.UserException;
+import com.example.tnote.base.exception.user.UserErrorResult;
+import com.example.tnote.base.exception.user.UserException;
 import com.example.tnote.base.utils.DateUtils;
 import com.example.tnote.base.utils.FileUploadUtils;
 import com.example.tnote.boundedContext.consultation.dto.ConsultationDeleteResponseDto;
