+++ src/main/java/com/example/tnote/boundedContext/subject/service/SubjectService.java
@@ -1,13 +1,13 @@
 package com.example.tnote.boundedContext.subject.service;
 
-import com.example.tnote.base.exception.ScheduleErrorResult;
-import com.example.tnote.base.exception.ScheduleException;
-import com.example.tnote.base.exception.SubjectsErrorResult;
-import com.example.tnote.base.exception.SubjectsException;
-import com.example.tnote.base.exception.TodoErrorResult;
-import com.example.tnote.base.exception.TodoException;
-import com.example.tnote.base.exception.UserErrorResult;
-import com.example.tnote.base.exception.UserException;
+import com.example.tnote.base.exception.schedule.ScheduleErrorResult;
+import com.example.tnote.base.exception.schedule.ScheduleException;
+import com.example.tnote.base.exception.subject.SubjectsErrorResult;
+import com.example.tnote.base.exception.subject.SubjectsException;
+import com.example.tnote.base.exception.todo.TodoErrorResult;
+import com.example.tnote.base.exception.todo.TodoException;
+import com.example.tnote.base.exception.user.UserErrorResult;
+import com.example.tnote.base.exception.user.UserException;
 import com.example.tnote.boundedContext.schedule.entity.ClassDay;
 import com.example.tnote.boundedContext.schedule.entity.Schedule;
 import com.example.tnote.boundedContext.schedule.repository.ScheduleRepository;
