+++ src/main/java/com/example/tnote/boundedContext/user/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.example.tnote.boundedContext.user.service;
 
-import com.example.tnote.base.exception.UserErrorResult;
-import com.example.tnote.base.exception.UserException;
+import com.example.tnote.base.exception.user.UserErrorResult;
+import com.example.tnote.base.exception.user.UserException;
 import com.example.tnote.base.utils.JwtTokenProvider;
 import com.example.tnote.boundedContext.RefreshToken.entity.RefreshToken;
 import com.example.tnote.boundedContext.RefreshToken.service.RefreshTokenService;
